requirements.txt
anthropic
requests
openpyxl
pandas
scikit-learn
jinja2
python-dotenv
google-api-python-client
google-auth